Splits Union assertions and test classes into multiple files.
